CCM-15676: more cleanup after merge
