+++ lambdas/nhsapp-status-handler/package.json
@@ -3,8 +3,7 @@
     "@aws-sdk/client-dynamodb": "^3.981.0",
     "@aws-sdk/lib-dynamodb": "^3.908.0",
     "digital-letters-events": "^0.0.1",
-    "utils": "^0.0.1",
-    "zod": "^4.3.6"
+    "utils": "^0.0.1"
   },
   "devDependencies": {
     "@tsconfig/node22": "^22.0.2",
