+++ tests/pact-tests/package.json
@@ -5,6 +5,7 @@
     "@types/jest": "^29.5.14",
     "jest": "^29.7.0",
     "rimraf": "^6.0.1",
+    "utils": "^0.0.1",
     "zod": "^4.3.6"
   },
   "name": "nhs-notify-digital-letters-pact-tests",
