+++ utils/utils/src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './channel-status-published-event';
 export * from './pdm-types';
 export * from './sender';
 export * from './ttl-dynamodb-record';
